First Commit
This the Vending Machine code, which ask user to  turn on the machine if select the produt from the list, then it ask user to insert coin / bill to purchase the product. If the input amount is more then desire product price then return the change. Also it will make machine to default by removing money and increasing the product in machine.